changed quantitys to Double

+++ src/main/java/BatteryManagement/Battery.java
@@ -18,12 +18,12 @@ public class Battery extends StoreMedium {
     }
 
     @Override
-    public void fill(Object input, Integer quantity) {
+    public void fill(Object input, Double quantity) {
         super.fill(input, quantity);
     }
 
     @Override
-    public List<Object> remove(Integer quantity) {
+    public List<Object> remove(Double quantity) {
         return super.remove(quantity);
     }
 }
