+++ src/main/java/Store/IStoreMedium.java
@@ -3,7 +3,7 @@ package Store;
 import java.util.List;
 
 public interface IStoreMedium {
-    void fill(Object input, Integer quantity);
-    List<Object> remove(Integer quantity);
+    void fill(Object input, Double quantity);
+    List<Object> remove(Double quantity);
     Double getRelativeFillState();
 }
